layer2/analyzer/requirements.txt
faster-whisper
anthropic
redis
ffmpeg-python
pyyaml
